+++ src/main/java/pe/upc/finanzas/repository/IPeriodo_TasaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import pe.upc.finanzas.entity.Periodo_Tasa;
 
 @Repository
-public interface IPeriodo_TasaRepository extends JpaRepository<Periodo_Tasa, Integer>{
+public interface IPeriodo_TasaRepository extends JpaRepository<Periodo_Tasa, Long>{
 
 }
