+++ src/main/java/pe/upc/finanzas/repository/ITipo_TasaRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import pe.upc.finanzas.entity.Tipo_Tasa;
 
 @Repository
-public interface ITipo_TasaRepository extends JpaRepository<Tipo_Tasa, Integer> {
+public interface ITipo_TasaRepository extends JpaRepository<Tipo_Tasa, Long> {
 
 	List<Tipo_Tasa> findByNTipo_Tasa(String NTipo_Tasa) throws Exception;
