+++ src/main/java/pe/upc/finanzas/service/IClienteService.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import pe.upc.finanzas.entity.Cliente;
 
-public interface IClienteService extends CrudService<Cliente,Integer>{
+public interface IClienteService extends CrudService<Cliente,Long>{
 
 	Optional<Cliente> findByNumTelefono(int NumTelefono) throws Exception;
 	Optional<Cliente> findByNumDNI(String NumDNI) throws Exception;
