+++ src/main/java/pe/upc/finanzas/service/ISistemaService.java
@@ -2,6 +2,6 @@ package pe.upc.finanzas.service;
 
 import pe.upc.finanzas.entity.Sistema;
 
-public interface ISistemaService extends CrudService<Sistema, Integer> {
+public interface ISistemaService extends CrudService<Sistema, Long> {
 
 }
