+++ src/main/java/pe/upc/finanzas/service/ITipoMonedaService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import pe.upc.finanzas.entity.TipoMoneda;
 
-public interface ITipoMonedaService extends CrudService<TipoMoneda, Integer>{
+public interface ITipoMonedaService extends CrudService<TipoMoneda, Long>{
 
 	List<TipoMoneda> findByNTipoMoneda(String NTipoMoneda);
