+++ src/main/java/pe/upc/finanzas/service/ITipo_TasaService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import pe.upc.finanzas.entity.Tipo_Tasa;
 
-public interface ITipo_TasaService extends CrudService<Tipo_Tasa, Integer>{
+public interface ITipo_TasaService extends CrudService<Tipo_Tasa, Long>{
 
 	List<Tipo_Tasa> findByNTipo_Tasa(String NTipo_Tasa) throws Exception;
 }
